css-align-3_dev/html/place-self-shorthand-005.htm

<!DOCTYPE html>
<html><head><title>CSS Box Alignment: place-self shorthand - inherit value</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez">
<link href="http://www.w3.org/TR/css3-align/#place-self-property" rel="help">
<meta content="Check that place-self's 'inherit' value expands to 'align-self' and 'justify-self'." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        place-self: start end;
    }
</style>
</head><body><div id="log"></div>
<div id="test"></div>
<script>
    var child = document.createElement("div");
    document.getElementById("test").appendChild(child);
    child.setAttribute("style", "place-self: inherit");
    var style = getComputedStyle(child);

    test(function() {
        assert_equals(style.getPropertyValue("align-self"),
            "start", "place-self resolved value for align-self");
    }, "Check place-self: inherit - align-self resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-self"),
           "end", "place-self resolved value for justify-self");
    }, "Check place-self: inherit - justify-self resolved value");
</script>
</body></html>
